Fixed the refactor
