+++ service/src/main/java/com/codeforcommunity/ServiceMain.java
@@ -1,6 +1,8 @@
 package com.codeforcommunity;
 
+import com.codeforcommunity.api.authenticated.IProtectedBookLogProcessor;
 import com.codeforcommunity.api.authenticated.IProtectedCountryProcessor;
+import com.codeforcommunity.api.authenticated.IProtectedReportProcessor;
 import com.codeforcommunity.api.authenticated.IProtectedSchoolProcessor;
 import com.codeforcommunity.api.authenticated.IProtectedUserProcessor;
 import com.codeforcommunity.api.unauthenticated.IAuthProcessor;
@@ -8,7 +10,9 @@ import com.codeforcommunity.auth.JWTAuthorizer;
 import com.codeforcommunity.auth.JWTCreator;
 import com.codeforcommunity.auth.JWTHandler;
 import com.codeforcommunity.logger.SLogger;
+import com.codeforcommunity.processor.authenticated.ProtectedBookLogProcessorImpl;
 import com.codeforcommunity.processor.authenticated.ProtectedCountryProcessorImpl;
+import com.codeforcommunity.processor.authenticated.ProtectedReportProcessorImpl;
 import com.codeforcommunity.processor.authenticated.ProtectedSchoolProcessorImpl;
 import com.codeforcommunity.processor.authenticated.ProtectedUserProcessorImpl;
 import com.codeforcommunity.processor.unauthenticated.AuthProcessorImpl;
@@ -89,11 +93,14 @@ public class ServiceMain {
     IProtectedUserProcessor protectedUserProc = new ProtectedUserProcessorImpl(this.db, emailer);
     IProtectedCountryProcessor protectedCountryProc = new ProtectedCountryProcessorImpl(this.db);
     IProtectedSchoolProcessor protectedSchoolProc = new ProtectedSchoolProcessorImpl(this.db);
+    IProtectedReportProcessor protectedReportProcessor = new ProtectedReportProcessorImpl(this.db);
+    IProtectedBookLogProcessor protectedBookLogProcessor = new ProtectedBookLogProcessorImpl(this.db);
+
 
     // Create the API router and start the HTTP server
     ApiRouter router =
         new ApiRouter(
-            jwtAuthorizer, authProc, protectedUserProc, protectedCountryProc, protectedSchoolProc);
+            jwtAuthorizer, authProc, protectedUserProc, protectedCountryProc, protectedSchoolProc, protectedReportProcessor, protectedBookLogProcessor);
     startApiServer(router, vertx);
   }
