+++ service/src/main/java/com/codeforcommunity/processor/authenticated/ProtectedSchoolProcessorImpl.java
@@ -1,71 +1,47 @@
 package com.codeforcommunity.processor.authenticated;
 
-import static org.jooq.generated.Tables.BOOK_LOGS;
 import static org.jooq.generated.Tables.SCHOOLS;
 import static org.jooq.generated.Tables.SCHOOL_CONTACTS;
 import static org.jooq.generated.Tables.SCHOOL_REPORTS_WITHOUT_LIBRARIES;
 import static org.jooq.generated.Tables.SCHOOL_REPORTS_WITH_LIBRARIES;
-import static org.jooq.generated.Tables.USERS;
 
 import com.codeforcommunity.api.authenticated.IProtectedSchoolProcessor;
 import com.codeforcommunity.auth.JWTData;
-import com.codeforcommunity.dto.CsvSerializer;
-import com.codeforcommunity.dto.report.ReportGeneric;
-import com.codeforcommunity.dto.report.ReportGenericListResponse;
 import com.codeforcommunity.dto.report.ReportWithLibrary;
 import com.codeforcommunity.dto.report.ReportWithoutLibrary;
-import com.codeforcommunity.dto.report.UpsertReportWithLibrary;
-import com.codeforcommunity.dto.report.UpsertReportWithoutLibrary;
-import com.codeforcommunity.dto.school.BookLog;
-import com.codeforcommunity.dto.school.BookLogListResponse;
 import com.codeforcommunity.dto.school.School;
 import com.codeforcommunity.dto.school.SchoolContact;
 import com.codeforcommunity.dto.school.SchoolContactListResponse;
 import com.codeforcommunity.dto.school.SchoolListResponse;
 import com.codeforcommunity.dto.school.SchoolSummary;
-import com.codeforcommunity.dto.school.UpsertBookLogRequest;
 import com.codeforcommunity.dto.school.UpsertSchoolContactRequest;
 import com.codeforcommunity.dto.school.UpsertSchoolRequest;
 import com.codeforcommunity.enums.ContactType;
 import com.codeforcommunity.enums.Country;
-import com.codeforcommunity.enums.Grade;
 import com.codeforcommunity.enums.LibraryStatus;
 import com.codeforcommunity.exceptions.AdminOnlyRouteException;
-import com.codeforcommunity.exceptions.BookLogDoesNotExistException;
-import com.codeforcommunity.exceptions.CsvSerializerException;
-import com.codeforcommunity.exceptions.InvalidShipmentYearException;
-import com.codeforcommunity.exceptions.MalformedParameterException;
-import com.codeforcommunity.exceptions.NoReportByIdFoundException;
-import com.codeforcommunity.exceptions.NoReportFoundException;
 import com.codeforcommunity.exceptions.SchoolAlreadyExistsException;
 import com.codeforcommunity.exceptions.SchoolContactAlreadyExistsException;
 import com.codeforcommunity.exceptions.SchoolContactDoesNotExistException;
 import com.codeforcommunity.exceptions.SchoolDoesNotExistException;
-import com.codeforcommunity.exceptions.UserDoesNotExistException;
-import com.codeforcommunity.logger.SLogger;
+import com.codeforcommunity.util.ProcessorUtility;
 import java.sql.Timestamp;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.jooq.DSLContext;
-import org.jooq.generated.tables.records.BookLogsRecord;
 import org.jooq.generated.tables.records.SchoolContactsRecord;
-import org.jooq.generated.tables.records.SchoolReportsWithLibrariesRecord;
-import org.jooq.generated.tables.records.SchoolReportsWithoutLibrariesRecord;
 import org.jooq.generated.tables.records.SchoolsRecord;
-import org.jooq.generated.tables.records.UsersRecord;
-
 public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
 
-  private final SLogger logger = new SLogger(ProtectedSchoolProcessorImpl.class);
   private final DSLContext db;
+  private final ProcessorUtility util;
 
   public ProtectedSchoolProcessorImpl(DSLContext db) {
     this.db = db;
+    this.util = new ProcessorUtility(db);
   }
 
   @Override
@@ -149,7 +125,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
       school.store();
 
       Integer schoolId = school.getId();
-      List<SchoolContact> contacts = this.queryForSchoolContacts(schoolId);
+      List<SchoolContact> contacts = this.util.queryForSchoolContacts(schoolId);
 
       return new School(
           school.getId(),
@@ -170,7 +146,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
 
   @Override
   public SchoolContactListResponse getAllSchoolContacts(JWTData userData, int schoolId) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -186,7 +162,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
 
   @Override
   public SchoolContact getSchoolContact(JWTData userData, int schoolId, int contactId) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -208,7 +184,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
   @Override
   public SchoolContact createSchoolContact(
       JWTData userData, int schoolId, UpsertSchoolContactRequest upsertSchoolContactRequest) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -326,7 +302,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
   @Override
   public void updateSchool(
       JWTData userData, int schoolId, UpsertSchoolRequest upsertSchoolRequest) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -373,7 +349,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
       throw new AdminOnlyRouteException();
     }
 
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -388,7 +364,7 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
       throw new AdminOnlyRouteException();
     }
 
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
@@ -403,409 +379,13 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
       throw new AdminOnlyRouteException();
     }
 
-    SchoolsRecord school = this.queryForSchool(schoolId);
+    SchoolsRecord school = this.util.queryForSchool(schoolId);
     if (school == null) {
       throw new SchoolDoesNotExistException(schoolId);
     }
 
     school.setHidden(false);
     school.store();
-  }
-
-  @Override
-  public ReportWithLibrary createReportWithLibrary(
-      JWTData userData, int schoolId, UpsertReportWithLibrary req) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    // Update this school to have a new library status
-    school.setLibraryStatus(LibraryStatus.EXISTS);
-    school.store();
-
-    // Save a record to the school_reports_with_libraries table
-    SchoolReportsWithLibrariesRecord newReport = db.newRecord(SCHOOL_REPORTS_WITH_LIBRARIES);
-    storeReportWithLibrary(userData, schoolId, req, newReport);
-    newReport.refresh();
-
-    return new ReportWithLibrary(
-        newReport.getId(),
-        newReport.getCreatedAt(),
-        newReport.getUpdatedAt(),
-        newReport.getSchoolId(),
-        userData.getUserId(),
-        newReport.getNumberOfChildren(),
-        newReport.getNumberOfBooks(),
-        newReport.getMostRecentShipmentYear(),
-        newReport.getIsSharedSpace(),
-        newReport.getHasInvitingSpace(),
-        newReport.getAssignedPersonRole(),
-        newReport.getAssignedPersonTitle(),
-        newReport.getApprenticeshipProgram(),
-        newReport.getTrainsAndMentorsApprentices(),
-        newReport.getHasCheckInTimetables(),
-        newReport.getHasBookCheckoutSystem(),
-        newReport.getNumberOfStudentLibrarians(),
-        newReport.getReasonNoStudentLibrarians(),
-        newReport.getHasSufficientTraining(),
-        newReport.getTeacherSupport(),
-        newReport.getParentSupport(),
-        newReport.getVisitReason(),
-        newReport.getActionPlan(),
-        newReport.getSuccessStories(),
-        req.getGradesAttended(),
-        req.getCheckInTimetable(),
-        getUserName(userData.getUserId()),
-        getSchoolName(schoolId),
-        req.getCheckOutTimetable(),
-        req.getNumberOfStudentLibrariansTrainers());
-  }
-
-  @Override
-  public void updateReportWithLibrary(
-      JWTData userData, int schoolId, int reportId, UpsertReportWithLibrary req) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    // Save a record to the school_reports_with_libraries table
-    SchoolReportsWithLibrariesRecord newReport =
-        db.selectFrom(SCHOOL_REPORTS_WITH_LIBRARIES)
-            .where(SCHOOL_REPORTS_WITH_LIBRARIES.ID.eq(reportId))
-            .fetchOne();
-
-    if (newReport == null) {
-      throw new NoReportFoundException(schoolId);
-    }
-
-    if (!userData.isAdmin() && !newReport.getUserId().equals(userData.getUserId())) {
-      throw new AdminOnlyRouteException();
-    }
-
-
-    if (req.getMostRecentShipmentYear() != null) {
-      if (isShipmentYearInvalid(req.getMostRecentShipmentYear())) {
-      throw new InvalidShipmentYearException(req.getMostRecentShipmentYear());
-      }
-    }
-
-    storeReportWithLibrary(userData, schoolId, req, newReport);
-  }
-
-  private void storeReportWithLibrary(
-      JWTData userData,
-      int schoolId,
-      UpsertReportWithLibrary req,
-      SchoolReportsWithLibrariesRecord newReport) {
-    newReport.setUserId(userData.getUserId());
-    newReport.setSchoolId(schoolId);
-    newReport.setNumberOfChildren(req.getNumberOfChildren());
-    newReport.setNumberOfBooks(req.getNumberOfBooks());
-    newReport.setMostRecentShipmentYear(req.getMostRecentShipmentYear());
-    newReport.setIsSharedSpace(req.getIsSharedSpace());
-    newReport.setHasInvitingSpace(req.getHasInvitingSpace());
-    newReport.setAssignedPersonRole(req.getAssignedPersonRole());
-    newReport.setAssignedPersonTitle(req.getAssignedPersonTitle());
-    newReport.setApprenticeshipProgram(req.getApprenticeshipProgram());
-    newReport.setTrainsAndMentorsApprentices(req.getTrainsAndMentorsApprentices());
-    newReport.setHasCheckInTimetables(req.getHasCheckInTimetables());
-    newReport.setHasBookCheckoutSystem(req.getHasBookCheckoutSystem());
-    newReport.setNumberOfStudentLibrarians(req.getNumberOfStudentLibrarians());
-    newReport.setReasonNoStudentLibrarians(req.getReasonNoStudentLibrarians());
-    newReport.setHasSufficientTraining(req.getHasSufficientTraining());
-    newReport.setTeacherSupport(req.getTeacherSupport());
-    newReport.setParentSupport(req.getParentSupport());
-    newReport.setVisitReason(req.getVisitReason());
-    newReport.setActionPlan(req.getActionPlan());
-    newReport.setSuccessStories(req.getSuccessStories());
-    newReport.setGradesAttended(
-        (Object[]) req.getGradesAttended().stream().map(Grade::name).toArray(String[]::new));
-
-    newReport.setCheckinTimetable(req.getCheckInTimetable().toString());
-    newReport.setCheckoutTimetable(req.getCheckOutTimetable().toString());
-    newReport.setNumberOfStudentLibrariansTrainers(req.getNumberOfStudentLibrariansTrainers());
-    newReport.store();
-  }
-
-  @Override
-  public ReportGeneric getMostRecentReport(JWTData userData, int schoolId) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    ReportGeneric report = null;
-    LibraryStatus libraryStatus = school.getLibraryStatus();
-
-    if (libraryStatus == LibraryStatus.EXISTS) {
-      report =
-          ReportWithLibrary.instantiateFromRecord(
-              db.selectFrom(SCHOOL_REPORTS_WITH_LIBRARIES)
-                  .where(SCHOOL_REPORTS_WITH_LIBRARIES.DELETED_AT.isNull())
-                  .and(SCHOOL_REPORTS_WITH_LIBRARIES.SCHOOL_ID.eq(schoolId))
-                  .orderBy(SCHOOL_REPORTS_WITH_LIBRARIES.ID.desc())
-                  .limit(1)
-                  .fetchOne(),
-              getUserName(userData.getUserId()),
-              getSchoolName(schoolId));
-    } else if (libraryStatus == LibraryStatus.DOES_NOT_EXIST) {
-      report =
-          ReportWithoutLibrary.instantiateFromRecord(
-              db.selectFrom(SCHOOL_REPORTS_WITHOUT_LIBRARIES)
-                  .where(SCHOOL_REPORTS_WITHOUT_LIBRARIES.DELETED_AT.isNull())
-                  .and(SCHOOL_REPORTS_WITHOUT_LIBRARIES.SCHOOL_ID.eq(schoolId))
-                  .orderBy(SCHOOL_REPORTS_WITHOUT_LIBRARIES.ID.desc())
-                  .limit(1)
-                  .fetchOne(),
-              getUserName(userData.getUserId()),
-              getSchoolName(schoolId));
-    }
-
-    if (report == null) {
-      logger.error(String.format("Report was not found for school with ID: %d", schoolId));
-      throw new NoReportFoundException(schoolId);
-    }
-
-    return report;
-  }
-
-  @Override
-  public ReportWithoutLibrary createReportWithoutLibrary(
-      JWTData userData, int schoolId, UpsertReportWithoutLibrary req) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    if (req.getMostRecentShipmentYear() != null) {
-      if (isShipmentYearInvalid(req.getMostRecentShipmentYear())) {
-        throw new InvalidShipmentYearException(req.getMostRecentShipmentYear());
-      }
-    }
-    school.setLibraryStatus(LibraryStatus.DOES_NOT_EXIST);
-    school.store();
-
-    SchoolReportsWithoutLibrariesRecord newReport = db.newRecord(SCHOOL_REPORTS_WITHOUT_LIBRARIES);
-    storeReportWithoutLibrary(userData, schoolId, req, newReport, req.getGradesAttended());
-    newReport.refresh();
-
-    return new ReportWithoutLibrary(
-        newReport.getId(),
-        newReport.getCreatedAt(),
-        newReport.getUpdatedAt(),
-        newReport.getSchoolId(),
-        newReport.getUserId(),
-        newReport.getNumberOfChildren(),
-        newReport.getNumberOfBooks(),
-        newReport.getMostRecentShipmentYear(),
-        newReport.getWantsLibrary(),
-        newReport.getHasSpace(),
-        req.getCurrentStatus(),
-        newReport.getReasonWhyNot(),
-        newReport.getReadyTimeline(),
-        newReport.getVisitReason(),
-        newReport.getActionPlan(),
-        newReport.getSuccessStories(),
-        req.getGradesAttended(),
-        getUserName(userData.getUserId()),
-        getSchoolName(schoolId),
-        newReport.getReasonNoLibrarySpace()
-        );
-  }
-
-  @Override
-  public void updateReportWithoutLibrary(
-      JWTData userData, int schoolId, int reportId, UpsertReportWithoutLibrary req) {
-
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    // Save a record to the school_reports_with_libraries table
-    SchoolReportsWithoutLibrariesRecord newReport =
-        db.selectFrom(SCHOOL_REPORTS_WITHOUT_LIBRARIES)
-            .where(SCHOOL_REPORTS_WITHOUT_LIBRARIES.ID.eq(reportId))
-            .fetchOne();
-
-    if (newReport == null) {
-      throw new NoReportFoundException(schoolId);
-    }
-
-    if (!userData.isAdmin() && !newReport.getUserId().equals(userData.getUserId())) {
-      throw new AdminOnlyRouteException();
-    }
-
-    if (req.getMostRecentShipmentYear() != null) {
-      if (isShipmentYearInvalid(req.getMostRecentShipmentYear())) {
-        throw new InvalidShipmentYearException(req.getMostRecentShipmentYear());
-      }
-    }
-
-    storeReportWithoutLibrary(userData, schoolId, req, newReport, req.getGradesAttended());
-  }
-
-  private void storeReportWithoutLibrary(
-      JWTData userData,
-      int schoolId,
-      UpsertReportWithoutLibrary req,
-      SchoolReportsWithoutLibrariesRecord newReport,
-      List<Grade> gradesAttended) {
-    newReport.setSchoolId(schoolId);
-    newReport.setUserId(userData.getUserId());
-    newReport.setNumberOfChildren(req.getNumberOfChildren());
-    newReport.setNumberOfBooks(req.getNumberOfBooks());
-    newReport.setMostRecentShipmentYear(req.getMostRecentShipmentYear());
-    newReport.setHasSpace(req.getHasSpace());
-    newReport.setCurrentStatus((Object[]) req.getCurrentStatus().toArray(new String[0]));
-    newReport.setReasonWhyNot(req.getReason());
-    newReport.setWantsLibrary(req.getWantsLibrary());
-    newReport.setReadyTimeline(req.getReadyTimeline());
-    newReport.setVisitReason(req.getVisitReason());
-    newReport.setActionPlan(req.getActionPlan());
-    newReport.setSuccessStories(req.getSuccessStories());
-    newReport.setGradesAttended(
-        (Object[]) gradesAttended.stream().map(Grade::name).toArray(String[]::new));
-    newReport.setReasonNoLibrarySpace(req.getReasonNoLibrarySpace());
-    newReport.store();
-  }
-
-  @Override
-  public ReportGenericListResponse getPaginatedReports(JWTData userData, int schoolId, int page) {
-    if (page < 1) {
-      throw new MalformedParameterException("p");
-    }
-
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    List<ReportWithLibrary> withLibraryReports =
-        db.selectFrom(SCHOOL_REPORTS_WITH_LIBRARIES)
-            .where(SCHOOL_REPORTS_WITH_LIBRARIES.DELETED_AT.isNull())
-            .and(SCHOOL_REPORTS_WITH_LIBRARIES.SCHOOL_ID.eq(schoolId)).fetch().stream()
-            .map(
-                record ->
-                    ReportWithLibrary.instantiateFromRecord(
-                        record, getUserName(userData.getUserId()), getSchoolName(schoolId)))
-            .collect(Collectors.toList());
-
-    List<ReportWithoutLibrary> noLibraryReports =
-        db.selectFrom(SCHOOL_REPORTS_WITHOUT_LIBRARIES)
-            .where(SCHOOL_REPORTS_WITHOUT_LIBRARIES.DELETED_AT.isNull())
-            .and(SCHOOL_REPORTS_WITHOUT_LIBRARIES.SCHOOL_ID.eq(schoolId)).fetch().stream()
-            .map(
-                record ->
-                    ReportWithoutLibrary.instantiateFromRecord(
-                        record, getUserName(userData.getUserId()), getSchoolName(schoolId)))
-            .collect(Collectors.toList());
-
-    int countWithLibrary =
-        db.fetchCount(
-            db.selectFrom(SCHOOL_REPORTS_WITH_LIBRARIES)
-                .where(SCHOOL_REPORTS_WITH_LIBRARIES.DELETED_AT.isNull())
-                .and(SCHOOL_REPORTS_WITH_LIBRARIES.SCHOOL_ID.eq(schoolId)));
-
-    int countWithoutLibrary =
-        db.fetchCount(
-            db.selectFrom(SCHOOL_REPORTS_WITHOUT_LIBRARIES)
-                .where(SCHOOL_REPORTS_WITHOUT_LIBRARIES.DELETED_AT.isNull())
-                .and(SCHOOL_REPORTS_WITHOUT_LIBRARIES.SCHOOL_ID.eq(schoolId)));
-
-    List<ReportGeneric> reports = new ArrayList<>();
-    reports.addAll(withLibraryReports);
-    reports.addAll(noLibraryReports);
-    reports.sort(Comparator.comparing(ReportGeneric::getCreatedAt));
-
-    int maxCountPerPage = 10;
-    int from = (page - 1) * maxCountPerPage;
-    int to = Math.min(page * maxCountPerPage, reports.size());
-    List<ReportGeneric> paginatedReports =
-        (from >= reports.size()) ? new ArrayList<>() : reports.subList(from, to);
-
-    return new ReportGenericListResponse(
-        paginatedReports, (countWithLibrary + countWithoutLibrary));
-  }
-
-  @Override
-  public BookLog createBookLog(JWTData userData, int schoolId, UpsertBookLogRequest request) {
-    if (!userData.isAdmin()) {
-      throw new AdminOnlyRouteException();
-    }
-
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    Integer count = request.getCount();
-    Timestamp date = request.getDate();
-    String notes = request.getNotes();
-
-    BookLogsRecord log = db.newRecord(BOOK_LOGS);
-    log.setSchoolId(schoolId);
-    log.setCount(count);
-    log.setDate(date);
-    log.setNotes(notes);
-    log.store();
-
-    return new BookLog(log.getId(), log.getCount(), log.getDate(), log.getNotes());
-  }
-
-  @Override
-  public BookLog updateBookLog(
-      JWTData userData, int schoolId, int bookId, UpsertBookLogRequest request) {
-    if (!userData.isAdmin()) {
-      throw new AdminOnlyRouteException();
-    }
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    BookLogsRecord log =
-        db.selectFrom(BOOK_LOGS)
-            .where(BOOK_LOGS.DELETED_AT.isNull())
-            .and(BOOK_LOGS.ID.eq(bookId))
-            .fetchOne();
-
-    if (log == null) {
-      throw new BookLogDoesNotExistException(bookId);
-    }
-
-    Integer count = request.getCount();
-    Timestamp date = request.getDate();
-    String notes = request.getNotes();
-    log.setSchoolId(schoolId);
-    log.setCount(count);
-    log.setDate(date);
-    log.setNotes(notes);
-    log.store();
-
-    return new BookLog(log.getId(), log.getCount(), log.getDate(), log.getNotes());
-  }
-
-  @Override
-  public void deleteBookLog(JWTData userData, int schoolId, int bookId) {
-    if (!userData.isAdmin()) {
-      throw new AdminOnlyRouteException();
-    }
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    BookLogsRecord log = db.selectFrom(BOOK_LOGS).where(BOOK_LOGS.ID.eq(bookId)).fetchOne();
-
-    if (log == null) {
-      throw new BookLogDoesNotExistException(bookId);
-    }
-
-    log.setDeletedAt(Timestamp.from(Instant.now()));
-    log.store();
   }
 
   @Override
@@ -820,8 +400,8 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
                 record ->
                     ReportWithLibrary.instantiateFromRecord(
                         record,
-                        getUserName(record.getUserId()),
-                        getSchoolName(record.getSchoolId())))
+                        this.util.getUserName(record.getUserId()),
+                        this.util.getSchoolName(record.getSchoolId())))
             .collect(Collectors.toList());
 
     List<ReportWithoutLibrary> noLibraryReports =
@@ -832,8 +412,8 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
                 record ->
                     ReportWithoutLibrary.instantiateFromRecord(
                         record,
-                        getUserName(record.getUserId()),
-                        getSchoolName(record.getSchoolId())))
+                        this.util.getUserName(record.getUserId()),
+                        this.util.getSchoolName(record.getSchoolId())))
             .collect(Collectors.toList());
 
     for (ReportWithLibrary report : withLibraryReports) {
@@ -852,87 +432,5 @@ public class ProtectedSchoolProcessorImpl implements IProtectedSchoolProcessor {
             .fetchInto(SchoolSummary.class);
 
     return new SchoolListResponse(schools);
-  }
-
-  @Override
-  public BookLogListResponse getBookLog(JWTData userData, int schoolId) {
-    SchoolsRecord school = this.queryForSchool(schoolId);
-    if (school == null) {
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-
-    List<BookLog> logs =
-        db.selectFrom(BOOK_LOGS)
-            .where(BOOK_LOGS.SCHOOL_ID.eq(schoolId))
-            .and(BOOK_LOGS.DELETED_AT.isNull())
-            .fetchInto(BookLog.class);
-
-    return (logs != null)
-        ? new BookLogListResponse(logs)
-        : new BookLogListResponse(new ArrayList<>());
-  }
-
-  @Override
-  public String getReportAsCsv(JWTData userData, int reportId, boolean hasLibrary) {
-    ReportGeneric report;
-    if (hasLibrary) {
-      report =
-          db.selectFrom(SCHOOL_REPORTS_WITH_LIBRARIES)
-              .where(SCHOOL_REPORTS_WITH_LIBRARIES.ID.eq(reportId))
-              .fetchOneInto(ReportWithLibrary.class);
-    } else {
-      report =
-          db.selectFrom(SCHOOL_REPORTS_WITHOUT_LIBRARIES)
-              .where(SCHOOL_REPORTS_WITHOUT_LIBRARIES.ID.eq(reportId))
-              .fetchOneInto(ReportWithoutLibrary.class);
-    }
-    if (report == null) {
-      throw new NoReportByIdFoundException(reportId);
-    }
-    StringBuilder builder = new StringBuilder();
-    try {
-      builder.append(CsvSerializer.getObjectHeader(report));
-      builder.append(CsvSerializer.toCsv(report));
-    } catch (IllegalStateException e) {
-      throw new CsvSerializerException(reportId);
-    }
-
-    return builder.toString();
-  }
-
-  private SchoolsRecord queryForSchool(int schoolId) {
-    return db.selectFrom(SCHOOLS)
-        .where(SCHOOLS.ID.eq(schoolId))
-        .and(SCHOOLS.DELETED_AT.isNull())
-        .fetchOne();
-  }
-
-  private List<SchoolContact> queryForSchoolContacts(int schoolId) {
-    return db.selectFrom(SCHOOL_CONTACTS)
-        .where(SCHOOL_CONTACTS.DELETED_AT.isNull())
-        .and(SCHOOL_CONTACTS.SCHOOL_ID.eq(schoolId))
-        .fetchInto(SchoolContact.class);
-  }
-
-  private boolean isShipmentYearInvalid(Integer year) {
-    return year <= 999 || year >= 10000;
-  }
-
-  private String getUserName(int userId) {
-    UsersRecord userRecord = db.selectFrom(USERS).where(USERS.ID.eq(userId)).fetchOne();
-    if (userRecord == null) {
-      logger.error(String.format("No username found for userId:  %d", userId));
-      throw new UserDoesNotExistException(userId);
-    }
-    return userRecord.getFirstName() + " " + userRecord.getLastName();
-  }
-
-  private String getSchoolName(int schoolId) {
-    SchoolsRecord schoolRecord = db.selectFrom(SCHOOLS).where(SCHOOLS.ID.eq(schoolId)).fetchOne();
-    if (schoolRecord == null) {
-      logger.error(String.format("No school name found for schoolId:  %d", schoolId));
-      throw new SchoolDoesNotExistException(schoolId);
-    }
-    return schoolRecord.getName();
   }
 }
